+++ src/applications/debt-letters/components/DebtLetterCard.jsx
@@ -5,10 +5,10 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
 import { bindActionCreators } from 'redux';
-
 import { deductionCodes } from '../const/deduction-codes';
 import { setActiveDebt as setDebt } from '../actions';
 import { renderAdditionalInfo } from '../const/diary-codes';
+import { formatter } from '../utils/page';
 
 const DebtLetterCard = ({ debt, setActiveDebt }) => {
   // TODO: currently we do not have a debtID so we need to make one by combining fileNumber and diaryCode
@@ -16,12 +16,6 @@ const DebtLetterCard = ({ debt, setActiveDebt }) => {
 
   const debtCardHeading =
     deductionCodes[debt.deductionCode] || debt.benefitType;
-
-  const formatter = new Intl.NumberFormat('en-US', {
-    style: 'currency',
-    currency: 'USD',
-    minimumFractionDigits: 2,
-  });
 
   const additionalInfo = renderAdditionalInfo(
     debt.diaryCode,
