+++ src/applications/debt-letters/utils/page.js
@@ -15,3 +15,33 @@ export const setPageFocus = (selector = '.va-nav-breadcrumbs') => {
     setFocus('#main h1');
   }
 };
+
+export const formatter = new Intl.NumberFormat('en-US', {
+  style: 'currency',
+  currency: 'USD',
+  minimumFractionDigits: 2,
+});
+
+/* 
+  TODO TECH DEBT: https://github.com/department-of-veterans-affairs/va.gov-team/issues/27790
+  Once debt.id is available via backend and endpoint to fetch single debtById is created
+  remove getCurrentDebt and replace with backend single item call
+*/
+export const getCurrentDebt = (selectedDebt, debts, location) => {
+  if (Object.keys(selectedDebt).length !== 0) {
+    return selectedDebt;
+  }
+  // get debtId out of the URL
+  const urlDebtId = location.pathname.replace(/[^0-9]/g, '');
+  // create debtIds derived from the debt fileNumber and deductionCode and add to debts
+  const debtsWithId = debts.reduce((acc, debt) => {
+    acc.push({
+      ...debt,
+      debtId: `${debt.fileNumber + debt.deductionCode}`,
+    });
+    return acc;
+  }, []);
+  // return debt that has the same debtId as the currentDebt
+  const [filteredDebt] = debtsWithId.filter(debt => debt.debtId === urlDebtId);
+  return filteredDebt;
+};
